+++ metrics-tracing-context/tests/integration.rs
@@ -1,4 +1,4 @@
-use metrics::{counter, Key, KeyData, Label, SharedString};
+use metrics::{counter, Key, Label, SharedString};
 use metrics_tracing_context::{LabelFilter, MetricsLayer, TracingContextLayer};
 use metrics_util::{
     layers::Layer, CompositeKey, DebugValue, DebuggingRecorder, MetricKind, Snapshotter,
@@ -108,7 +108,7 @@ fn test_basic_functionality() {
         vec![(
             CompositeKey::new(
                 MetricKind::Counter,
-                Key::Owned(KeyData::from_static_parts(LOGIN_ATTEMPTS, SVC_USER_EMAIL))
+                Key::from_static_parts(LOGIN_ATTEMPTS, SVC_USER_EMAIL)
             ),
             None,
             None,
@@ -145,7 +145,7 @@ fn test_macro_forms() {
             (
                 CompositeKey::new(
                     MetricKind::Counter,
-                    Key::Owned(KeyData::from_static_parts(LOGIN_ATTEMPTS_NONE, USER_EMAIL))
+                    Key::from_static_parts(LOGIN_ATTEMPTS_NONE, USER_EMAIL)
                 ),
                 None,
                 None,
@@ -154,10 +154,7 @@ fn test_macro_forms() {
             (
                 CompositeKey::new(
                     MetricKind::Counter,
-                    Key::Owned(KeyData::from_static_parts(
-                        LOGIN_ATTEMPTS_STATIC,
-                        SVC_USER_EMAIL
-                    )),
+                    Key::from_static_parts(LOGIN_ATTEMPTS_STATIC, SVC_USER_EMAIL),
                 ),
                 None,
                 None,
@@ -166,10 +163,7 @@ fn test_macro_forms() {
             (
                 CompositeKey::new(
                     MetricKind::Counter,
-                    Key::Owned(KeyData::from_static_parts(
-                        LOGIN_ATTEMPTS_DYNAMIC,
-                        NODE_USER_EMAIL
-                    )),
+                    Key::from_static_parts(LOGIN_ATTEMPTS_DYNAMIC, NODE_USER_EMAIL),
                 ),
                 None,
                 None,
@@ -178,10 +172,7 @@ fn test_macro_forms() {
             (
                 CompositeKey::new(
                     MetricKind::Counter,
-                    Key::Owned(KeyData::from_static_parts(
-                        LOGIN_ATTEMPTS_BOTH,
-                        SVC_NODE_USER_EMAIL
-                    )),
+                    Key::from_static_parts(LOGIN_ATTEMPTS_BOTH, SVC_NODE_USER_EMAIL),
                 ),
                 None,
                 None,
@@ -205,10 +196,7 @@ fn test_no_labels() {
     assert_eq!(
         snapshot,
         vec![(
-            CompositeKey::new(
-                MetricKind::Counter,
-                Key::Owned(KeyData::from_static_name(LOGIN_ATTEMPTS)),
-            ),
+            CompositeKey::new(MetricKind::Counter, Key::from_static_name(LOGIN_ATTEMPTS),),
             None,
             None,
             DebugValue::Counter(1),
@@ -261,7 +249,7 @@ fn test_multiple_paths_to_the_same_callsite() {
             (
                 CompositeKey::new(
                     MetricKind::Counter,
-                    Key::Owned(KeyData::from_static_parts(MY_COUNTER, SAME_CALLSITE_PATH_1)),
+                    Key::from_static_parts(MY_COUNTER, SAME_CALLSITE_PATH_1),
                 ),
                 None,
                 None,
@@ -270,7 +258,7 @@ fn test_multiple_paths_to_the_same_callsite() {
             (
                 CompositeKey::new(
                     MetricKind::Counter,
-                    Key::Owned(KeyData::from_static_parts(MY_COUNTER, SAME_CALLSITE_PATH_2)),
+                    Key::from_static_parts(MY_COUNTER, SAME_CALLSITE_PATH_2),
                 ),
                 None,
                 None,
@@ -320,7 +308,7 @@ fn test_nested_spans() {
         vec![(
             CompositeKey::new(
                 MetricKind::Counter,
-                Key::Owned(KeyData::from_static_parts(MY_COUNTER, COMBINED_LABELS))
+                Key::from_static_parts(MY_COUNTER, COMBINED_LABELS)
             ),
             None,
             None,
@@ -356,7 +344,7 @@ fn test_label_filtering() {
         vec![(
             CompositeKey::new(
                 MetricKind::Counter,
-                Key::Owned(KeyData::from_static_parts(LOGIN_ATTEMPTS, EMAIL_USER))
+                Key::from_static_parts(LOGIN_ATTEMPTS, EMAIL_USER)
             ),
             None,
             None,