+++ metrics-util/benches/filter.rs
@@ -1,7 +1,7 @@
 use criterion::{criterion_group, criterion_main, Criterion};
 
 #[cfg(feature = "layer-filter")]
-use metrics::{Key, KeyData, Label, NoopRecorder, Recorder, SharedString};
+use metrics::{Key, Label, NoopRecorder, Recorder, SharedString};
 
 #[cfg(feature = "layer-filter")]
 use metrics_util::layers::{FilterLayer, Layer};
@@ -17,10 +17,10 @@ fn layer_benchmark(c: &mut Criterion) {
             let recorder = filter_layer.layer(NoopRecorder);
             static KEY_NAME: [SharedString; 1] = [SharedString::const_str("tokio.foo")];
             static KEY_LABELS: [Label; 1] = [Label::from_static_parts("foo", "bar")];
-            static KEY_DATA: KeyData = KeyData::from_static_parts(&KEY_NAME, &KEY_LABELS);
+            static KEY_DATA: Key = Key::from_static_parts(&KEY_NAME, &KEY_LABELS);
 
             b.iter(|| {
-                recorder.increment_counter(Key::Borrowed(&KEY_DATA), 1);
+                recorder.increment_counter(&KEY_DATA, 1);
             })
         });
         group.bench_function("no match", |b| {
@@ -29,10 +29,10 @@ fn layer_benchmark(c: &mut Criterion) {
             let recorder = filter_layer.layer(NoopRecorder);
             static KEY_NAME: [SharedString; 1] = [SharedString::const_str("hyper.foo")];
             static KEY_LABELS: [Label; 1] = [Label::from_static_parts("foo", "bar")];
-            static KEY_DATA: KeyData = KeyData::from_static_parts(&KEY_NAME, &KEY_LABELS);
+            static KEY_DATA: Key = Key::from_static_parts(&KEY_NAME, &KEY_LABELS);
 
             b.iter(|| {
-                recorder.increment_counter(Key::Borrowed(&KEY_DATA), 1);
+                recorder.increment_counter(&KEY_DATA, 1);
             })
         });
         group.bench_function("deep match", |b| {
@@ -44,20 +44,20 @@ fn layer_benchmark(c: &mut Criterion) {
                 SharedString::const_str("tokio.foo"),
             ];
             static KEY_LABELS: [Label; 1] = [Label::from_static_parts("foo", "bar")];
-            static KEY_DATA: KeyData = KeyData::from_static_parts(&KEY_NAME, &KEY_LABELS);
+            static KEY_DATA: Key = Key::from_static_parts(&KEY_NAME, &KEY_LABELS);
 
             b.iter(|| {
-                recorder.increment_counter(Key::Borrowed(&KEY_DATA), 1);
+                recorder.increment_counter(&KEY_DATA, 1);
             })
         });
         group.bench_function("noop recorder overhead (increment_counter)", |b| {
             let recorder = NoopRecorder;
             static KEY_NAME: [SharedString; 1] = [SharedString::const_str("tokio.foo")];
             static KEY_LABELS: [Label; 1] = [Label::from_static_parts("foo", "bar")];
-            static KEY_DATA: KeyData = KeyData::from_static_parts(&KEY_NAME, &KEY_LABELS);
+            static KEY_DATA: Key = Key::from_static_parts(&KEY_NAME, &KEY_LABELS);
 
             b.iter(|| {
-                recorder.increment_counter(Key::Borrowed(&KEY_DATA), 1);
+                recorder.increment_counter(&KEY_DATA, 1);
             })
         });
     }